Dedup manual JSON writer and memory CLI report via writeJsonArtifact
